[Java.Interop.BootstrapTasks] Use JdkInfo
Context: https://github.com/xamarin/xamarin-android-tools/pull/29#issuecomment-400048306

Bumps to xamarin/xamarin-android-tools:master@75530565

The `<JdkInfo/>` task within `Java.Interop.BootstrapTasks.dll` isn't
the *only* thing that needs to determine the path of an installed JDK.
The IDEs could also use such functionality, as could xamarin-android.

Add a submodule reference to xamarin-android-tools, and use
`Xamarin.Android.Tools.JdkInfo` to find the preferred JDK, instead of
having lots of custom probing with the `<JdkInfo/>` task.

Finally, for consistency, move `src/Java.Interop.BootstrapTasks` to
`build-tools/Java.Interop.BootstrapTasks`, as the
`Java.Interop.BootstrapTasks.dll` assembly is only intended for in
building Java.Interop *itself*, and not for "downlevel" distribution.
