+++ build-tools/Java.Interop.BootstrapTasks/Properties/AssemblyInfo.cs
